+++ mobile/android/android-components/components/feature/search/src/main/java/mozilla/components/feature/search/middleware/SearchMiddleware.kt
@@ -17,6 +17,8 @@ import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.SearchState
 import mozilla.components.feature.search.storage.BundledSearchEnginesStorage
 import mozilla.components.feature.search.storage.CustomSearchEngineStorage
+import mozilla.components.feature.search.storage.SearchEngineSelectorConfig
+import mozilla.components.feature.search.storage.SearchEngineSelectorRepository
 import mozilla.components.feature.search.storage.SearchMetadataStorage
 import mozilla.components.lib.state.Middleware
 import mozilla.components.lib.state.MiddlewareContext
@@ -60,11 +62,14 @@ class SearchMiddleware(
         context,
         additionalBundledSearchEngineIds.toSet(),
     ),
+    private val searchEngineSelectorConfig: SearchEngineSelectorConfig? = null,
     private val searchExtraParams: SearchExtraParams? = null,
     private val ioDispatcher: CoroutineContext = Dispatchers.IO,
 ) : Middleware<BrowserState, BrowserAction> {
     private val logger = Logger("SearchMiddleware")
     private val scope = CoroutineScope(ioDispatcher)
+    private val searchEngineSelectorRepository: SearchEngineRepository? =
+        searchEngineSelectorConfig?.let { SearchEngineSelectorRepository(it) }
 
     override fun invoke(
         context: MiddlewareContext<BrowserState, BrowserAction>,
@@ -107,14 +112,26 @@ class SearchMiddleware(
         val migrationValues = migration?.getValuesToMigrate()
         performCustomSearchEnginesMigration(migrationValues)
 
-        val regionBundle = async(ioDispatcher) {
-            bundleStorage.load(
-                region = region,
-                distribution = distribution,
-                searchExtraParams = searchExtraParams,
-                coroutineContext = ioDispatcher,
-            )
+        val regionBundle = if (searchEngineSelectorRepository != null) {
+            async(ioDispatcher) {
+                searchEngineSelectorRepository.load(
+                    region = region,
+                    distribution = distribution,
+                    searchExtraParams = searchExtraParams,
+                    coroutineContext = ioDispatcher,
+                )
+            }
+        } else {
+            async(ioDispatcher) {
+                bundleStorage.load(
+                    region = region,
+                    distribution = distribution,
+                    searchExtraParams = searchExtraParams,
+                    coroutineContext = ioDispatcher,
+                )
+            }
         }
+
         val customSearchEngines = async(ioDispatcher) { customStorage.loadSearchEngineList() }
         val hiddenSearchEngineIds = async(ioDispatcher) { metadataStorage.getHiddenSearchEngines() }
         val disabledSearchEngineIds = async(ioDispatcher) { metadataStorage.getDisabledSearchEngineIds() }
@@ -315,6 +332,25 @@ class SearchMiddleware(
             val list: List<SearchEngine>,
             val defaultSearchEngineId: String,
         )
+    }
+
+    /**
+     * A repository for search engines from Application Services.
+     */
+    interface SearchEngineRepository {
+        /**
+         * Loads the search engines for the given [locale] and [region] from Application Services.
+         *
+         * If [distribution] is not null then attempt to load the bundled search engine for the
+         * [distribution] in the specified [locale] and [region] if available.
+         */
+        suspend fun load(
+            region: RegionState,
+            locale: Locale = Locale.getDefault(),
+            distribution: String? = null,
+            searchExtraParams: SearchExtraParams? = null,
+            coroutineContext: CoroutineContext = Dispatchers.IO,
+        ): BundleStorage.Bundle
     }
 
     /**